% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{source_and_pb}
\alias{source_and_pb}
\alias{source_current_and_pb}
\alias{bg_and_pb}
\alias{bg_current_and_pb}
\alias{job_and_pb}
\alias{job_current_and_pb}
\title{Run Scripts and RStudio Jobs with Pushbullet alerts}
\usage{
source_and_pb(script_path = NULL, ...)

source_current_and_pb(...)

bg_and_pb(script_path = NULL)

bg_current_and_pb()

job_and_pb(script_path = NULL)

job_current_and_pb()
}
\arguments{
\item{script_path}{The script to run. If NULL for `job_and_pb()` and
`source_and_pb()`, prompts for a file interactively in the RStudio API.}

\item{...}{passed on to [source()]}
}
\description{
Functions, mostly to be used with RStudio Add-Ins, to run  scripts
with Pushbullet alerts. `source_` functions source in current session,
`bg_` functions in background session, and `job_. functions in the background
as RStudio jobs.   `_current_` functions use the current
RStudio editor pane, other functions take a script argument.

These functions will issue a push alert using [RPushbullet::pbPost()].
See [here](https://github.com/eddelbuettel/rpushbullet#initialization) about
setting up Pushbullet with R - you just need to get an API key and run
[RPushbullet::pbSetup()].
}
